requirements.txt
django==5.1
djangorestframework==3.15
django-cors-headers==4.4
requests==2.32.3